Phase 1 Step 5: extract OptionsService + PolygonGateway + options_math
